func/langchain/requirements.txt
langchain
langchain-openai
chainlit
python-dotenv